# routers/api/docs/swagger.yaml
basePath: /
definitions:
  Rsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/proto.CodeType'
        description: msg code
      message:
        description: msg info
        type: string
      requestId:
        description: trans id
        type: string
      results:
        description: response body
    type: object
  async.AsyncTask:
    properties:
      processed:
        type: integer
      taskId:
        type: string
      taskStatus:
        type: string
      total:
        type: integer
    type: object
...
  proto.CodeType:
    enum:
    - 200
    - 201
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    - 1021
    - 1022
    - 1027
    - 1028
    - 1029
    - 1030
    - 1031
    - 1032
    - 1036
    - 1037
    - 1038
    - 1046
    - 1050
    - 1061
    - 1062
    type: integer
    x-enum-comments:
      CodeCopyIdError: 文件夹不能拷贝到自己
      CodeFailedToConnect: 建立连接失败
      CodeFailedToConnectDB: 连接DB失败
      CodeFailedToCreateBucket: 失败去创建桶
      CodeFailedToCreateFolder: 失败去创建文件夹
      CodeFailedToCreateMultipartTask: 失败去创建任务
      CodeFailedToCreateSymlink: 失败去创建符号链接
      CodeFailedToDeleteUser: 删除用户失败
      CodeFailedToGetUsedStorage: 获取用户空间使用量失败
      CodeFailedToInitUser: 失败去初始化用户
      CodeFailedToOpenFile: 打开文件失败
      CodeFailedToOperateDB: 失败去操作数据库
      CodeFailedToOptMinio: 失败去写文件
      CodeFailedToSaveFile: 失败去保存数据
      CodeFileExist: 文件已存在
      CodeFileNotExist: 文件不存在
      CodeFolderDepthTooLong: 文件夹层数超过5层
      CodeFolderExist: 文件夹已存在
      CodeFolderNotExist: 文件夹不存在
      CodeGetAsyncTaskInfoFailed: 获取异步任务状态失败
      CodeMultipartRangeUploaded: 分片范围已上传
      CodeMultipartTaskCompleteErr: 合并错误
      CodeMultipartTaskHashErr: 文件hash校验错误
      CodeMultipartTaskNotFound: 未找到任务
      CodeMultipartTaskOverlap: 任务重叠
      CodeMultipartTaskRangeErr: 上传范围错误
      CodeMultipartUploadingConflit: 分片范围上传冲突
      CodeNotEnoughSpace: 空间不够，不上传
      CodeParamErr: 参数错误
      CodeReqParamErr: 请求参数错误(推荐)
      CodeUserIdError: 用户Id的错误
      CodeWriteRedisFailed: 写入redis 失败
      CodeZipFileFailed: 压缩文件失败
    x-enum-varnames:
    - CodeOk
    - CodeCreateAsyncTaskSuccess
    - CodeParamErr
    - CodeReqParamErr
    - CodeFileNotExist
    - CodeFailedToOpenFile
    - CodeFolderNotExist
    - CodeFailedToConnect
    - CodeFailedToConnectDB
    - CodeFailedToOptMinio
    - CodeFailedToCreateBucket
    - CodeFailedToSaveFile
    - CodeFailedToOperateDB
    - CodeFailedToCreateFolder
    - CodeFileExist
    - CodeFolderExist
    - CodeFolderDepthTooLong
    - CodeUserIdError
    - CodeFailedToInitUser
    - CodeFailedToDeleteUser
    - CodeFailedToGetUsedStorage
    - CodeCopyIdError
    - CodeFailedToCreateMultipartTask
    - CodeMultipartTaskNotFound
    - CodeMultipartTaskOverlap
    - CodeMultipartTaskRangeErr
    - CodeMultipartTaskHashErr
    - CodeMultipartTaskCompleteErr
    - CodeNotEnoughSpace
    - CodeMultipartRangeUploaded
    - CodeMultipartUploadingConflit
    - CodeWriteRedisFailed
    - CodeZipFileFailed
    - CodeFailedToCreateSymlink
    - CodeGetAsyncTaskInfoFailed
...
  proto.ErrMess:
    properties:
      code:
        $ref: '#/definitions/proto.CodeType'
      message:
        type: string
    type: object
...
info:
  contact:
    email: service@ao.space
    name: AO.space
    url: https://ao.space/
  description: This is AO.space aofs OpenAPI reference document.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: aofs apis
  version: "1.0"
paths:
  /space/v1/api/async/task:
    get:
      consumes:
      - application/json
      description: Query the status of an asynchronous task
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: task id
        in: query
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: response
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/async.AsyncTask'
              type: object
      tags:
      - Async
  /space/v1/api/file/compressed:
    get:
      description: Get compressed graph
      parameters:
      - description: uuid
        in: query
        name: uuid
        required: true
        type: string
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: file content
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.ErrMess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.ErrMess'
      summary: Get compressed graph
      tags:
      - File
...
  /space/v1/api/file/download:
    get:
      description: File download
      parameters:
      - description: uuid
        in: query
        name: uuid
        required: true
        type: string
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: range, such as：bytes=200-1000
        in: header
        name: Range
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: file content
          schema:
            type: file
        "206":
          description: Partial Content
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.ErrMess'
        "416":
          description: Range Not Satisfiable
          schema:
            $ref: '#/definitions/proto.ErrMess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.ErrMess'
      summary: File download
      tags:
      - File
...
  /space/v1/api/file/thumb:
    get:
      description: Get thumbnail
      parameters:
      - description: uuid
        in: query
        name: uuid
        required: true
        type: string
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: file content
          schema:
            type: file
        "400":
          description: param error
          schema:
            $ref: '#/definitions/proto.ErrMess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.ErrMess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.ErrMess'
      summary: Get thumbnail
      tags:
      - File
...
swagger: "2.0"